ppt_mp3_integration/requirements.txt
#!/usr/bin/env python3
"""
Requirements file for PowerPoint MP3 Integration Project
"""

Flask==2.0.1
Werkzeug==2.0.1
python-pptx==0.6.21
pydub==0.25.1
SpeechRecognition==3.8.1
gunicorn==20.1.0
pytest==6.2.5
requests==2.26.0
